# docs/swagger.yaml
openapi: "3.0.0"
servers:
  - url: 'https://ubademy-14-prod.herokuapp.com/'
info:
  description: Exams API encargada de todas las operaciones relacionadas a examenes
  version: 0.2.0
  title: Exams

tags:
  - name: Exam
    description: Operaciones sobre examenes

paths:
  '/exams/create':
    'post':
      tags:
        - Exam
      description: Crea un examen nuevo
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que desea crear el examen
      requestBody:
        description: body para la creacion del examen
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/examCreate'
      responses:
        '200':
          description: Examen creado correctamente
        '400':
          description: La pregunta o el título no pueden estar vacíos/son inválidos
        '403':
          description: Invalid user action

  '/exams/edit/{exam_id}':
    'put':
      tags:
        - Exam
      description: Edita un examen que aun no fue publicado
      parameters:
        - in: path
          required: true
          name: exam_id
          description: id del examen
          example: 5
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que desea crear el examen
      requestBody:
        description: body para la edicion del examen
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/examCreate'
      responses:
        '200':
          description: Examen editado correctamente
        '400':
          description: Bad request
        '403':
          description: Invalid user action

  '/exams/publish':
    'post':
      tags:
        - Exam
      description: Publica un examen que está como borrador
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que desea publicar el examen
      requestBody:
        description: body para la publicacion del examen
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/examPublish'
      responses:
        '200':
          description: Examen publicado correctamente
        '400':
          description: Bad request
        '403':
          description: Invalid user action

  '/exams/{course_id}':
    'get':
      tags:
        - Exam
      description: Obtengo una lista con todos los examenes del curso en caso de ser el creador, y todos los que no están publicados si soy alumno
      parameters:
        - in: path
          required: true
          name: course_id
          description: id del curso
          example: 5
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario

      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/infoExamsToCorrect'
        '204':
          description: No content
        '400':
          description: Bad Request
        '403':
          description: Invalid user action
  
  '/exams/correct/{exam_id}/{student_id}':
    'get':
      tags:
        - Exam
      description: Obtengo detalle de un examen sin corregir para corregirlo
      parameters:
        - in: path
          name: exam_id
          description: id del examen
          required: true
        - in: path
          required: true
          name: student_id
          description: id del estudiante
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario

      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/infoExamToCorrect'
        '204':
          description: No content
        '400':
          description: Bad Request
        '403':
          description: Invalid user action
        '404':
          description: Course or student not found
    'patch':
      tags:
        - Exam
      description: Envio correcciones de un examen
      parameters:
        - in: path
          name: exam_id
          description: id del examen
          required: true
        - in: path
          name: student_id
          description: id del alumno que realizo el examen
          required: true
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      requestBody:
        description: Body para corregir el examen
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/infoExamCorrected'

      responses:
        '204':
          description: No content
        '400':
          description: Bad Request, campos invalidos
        '403':
          description: Invalid user action
        '404':
          description: Course or student not found

  '/exams/view/{course_id}':
    'get':
      tags:
        - Exam
      description: Obtiene la lista de los examenes que el alumno realizo
      parameters:
        - in: path
          name: course_id
          description: id del curso
          required: true
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/toDoExamsInfo'
        '204':
          description: No content
        '404':
          description: Curso no existe
        '400':
          description: Bad Request
        '403':
          description: Invalid user action 

  '/exams/complete/{exam_id}':
    'get':
      tags:
        - Exam
      description: Obtiene la información del examen indicado por el id, para ser completado por el alumno
      parameters:
        - in: path
          required: true
          name: exam_id
          description: id del examen
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/toDoExamInfo'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Curso no existe
        '403':
          description: Invalid user action 

    'post':
      tags:
        - Exam
      description: Completo un examen
      parameters:
        - in: path
          required: true
          name: exam_id
          description: id del examen
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      requestBody:
        description: Body para completar el examen
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/examCompletionInfo'
      responses:
        '200':
          description: Todo okay
        '400':
          description: Bad Request
        '404':
          description: Curso no existe
        '403':
          description: Invalid user action 

components:
  schemas:
    examCreate:
      type: object
      required:
        - id_course
        - name
        - questions
      properties:
        id_course:
          type: integer
          example: 8
        name:
          type: string
          example: 'Examen piloto sobre conocimientos de python3'
        questions:
          type: array
          items:
            type: string
          description: Un array con las preguntas que va a tener el examen
          example: ['¿Existe la expresión i++ en python?', 'Nombre alguna funcionalidad que este en Python3 y no en Python 2.7']

    infoExamsToCorrect:
      type: object
      properties:
        exam_name:
          type: string
          example: 'Examen de conocimiento general'
        id_student:
          type: integer
          example: 5
        questions:
          type: array
          items:
            type: string
          description: Un array con las preguntas que tiene el examen 
          example: ['¿Quien descubrio america?', '¿Cuanto es 2+2?']
        answers:
          type: array
          items:
            type: string
          description: Un array con las respuestas del alumno
          example: ['Colón', '4']
  
    infoExamToCorrect:
      type: object
      properties:
        id_course:
          type: integer
          example: 8
        id_exam:
          type: integer
          example: 8
        id_student:
          type: integer
          example: 8
        questions:
          type: array
          items:
            type: string
          description: un array con las preguntas del examen con el orden
          example: ['¿Quien descubrio america?', '¿Cuanto es 2+2?']
        answers:
          type: array
          items:
            type: string  
          description: un array con las respuestas realizadas por el alumno, con el orden
          example: ["Carlos Saul Menem", "85"]
        status:
          type: string
          description: estado del examen, not corrected
          example: not corrected   

    infoExamCorrected:
      type: object
      properties:
        id_course:
          type: integer
          example: 8
        id_exam:
          type: integer
          example: 8
        id_student:
          type: integer
          example: 8
        corrections:
          type: string
          description: la correccion realizada 
          example: 'El examen en temrinos generales, esta bastante bien, se podria haber desarrollado un poco mas en la ultima pregunta'
        status:
          type: string
          description: estado del examen, las opciones son failed o passed
          example: 'failed'   

    toDoExamsInfo:
      type: object
      properties:
        name:
          type: string
          example: 'Primer examen del curso de jardineria'
        id_exam:
          type: integer
          example: 1
        id_course:
          type: integer
          description: id del curso
          example: 5
        questions: 
          type: array
          items:
            type: string
          description: Un array con las preguntas del examen
          example: ["¿Qué es una planta?", "Describa tres metodos de jardineria adquiridos a lo largo del curso"]
    
    examPublish:
      type: object
      properties:
        exam_id:
          type: integer
          description: Id del examen a ser publicado
          example: 5
    
    toDoExamInfo:
      type: object
      properties:
        name:
          type: string
          example: 'Primer examen del curso de jardineria'
        id_exam:
          type: integer
          example: 1
        questions: 
          type: array
          items:
            type: string
          description: Un array con las preguntas del examen
          example: ["¿Qué es una planta?", "Describa tres metodos de jardineria adquiridos a lo largo del curso"]
    
    examCompletionInfo:
      type: object
      properties:
        id_exam:
          type: integer
          example: 1
        id_student:
          type: integer
          example: 3
        questions: 
          type: array
          items:
            type: string
          description: Un array con las preguntas del examen
          example: ["¿Qué es una planta?", "Describa tres metodos de jardineria adquiridos a lo largo del curso"]
        answers:
          type: array
          items:
            type: string
          description: un array con las respuestas del alumno, también manteniendo el orden
          example: ["Algo verde que crece en el jardin", "No me acuerdo de ninguno"]